==> message-categorization/rnn-softmax-multi-class-text-classify/1-preprocess-clean.ipynb <==
# Overview

We'll be shaping the clean data.

# Instantiate required Python components.

Our project will use TensorFlow for developing our model.  We'll also need several other Python libraries to work with our CSV.

import re
import pandas as pd
import csv
import numpy as np

# Used for Troubleshooting
from IPython.display import display

# Set Hyperparameters

Merge multiple files into one CSV.  Ensure that the columns are in the following order: reason, singleMessage

# The file that contains the data.
FILE_MESSAGES = "./data/sources/20230328-clean-data.csv"

# Read the CSV Data

Read the CSV contents and keep only specific fields.

# Open file and save to dataframe.
df = pd.read_csv(FILE_MESSAGES)

#print(df.columns)
display(df)

# Preprocess Data

As part of the Machine Learning process, we will remove fields not required, fix missing values, remove noisy data, and any additional steps to prepare for the ML training process.

## Remove Empty Messages

# Remove rows that have empty data (missing values) in the specified column
df = df[df["message"].notna()]

## Keep Labels and Messages

We will keep only specific columns that is important to the model.

# Keep specific columns.
df = df[["message"]]

print(df.columns)
print(f'Total number of rows: {len(df)}')

## Add Reason Column

Our two features we want to keep are:

- reason
- singleMessage

We need to aim to format our datashape to that.

# Add a new column called 'reason' with a single value for all rows
df["reason"] = "Clean"

## Rename Column Names

df = df.rename(columns={"message": "singleMessage"})

## Change Order of Columns

df = df[["reason", "singleMessage"]]

# Review Data Results

display(df)

# 🚧 Save Data to Disk

Let's save all our hard work formatting the dataframe to a CSV for future reference.

- [Pandas DataFrame.to_csv](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.to_csv.html)

df.to_csv('data/output/1-preprocess-clean.csv', index=False)
