==> message-categorization/rnn-softmax-multi-class-text-classify/3-merge-data.ipynb <==
# Overview

We'll merge all extra data sets into one CSV.

# Instantiate required Python components.

Our project will use TensorFlow for developing our model.  We'll also need several other Python libraries to work with our CSV.

import re
import pandas as pd
import csv
import numpy as np

# Used for Troubleshooting
from IPython.display import display

# Set Hyperparameters

Merge multiple files into one CSV.  Ensure that the columns are in the following order: reason, singleMessage

# All the files to merge.
FILE_MESSAGES = [
    "./data/output/1-preprocessed-data.csv",
    "./data/output/1-preprocess-clean.csv",
    "./data/output/2-dataset-politics.csv"
]

# Read the CSV Data

# Specify the correct columns
correct_columns = ['reason', 'singleMessage']

# Initialize an empty list to store valid DataFrames
valid_dataframes = []

# Read each CSV file and check if it contains the correct columns
for file in FILE_MESSAGES:
    df = pd.read_csv(file)
    
    if set(df.columns) == set(correct_columns):
        valid_dataframes.append(df)
        print(f"{file} contains the correct columns.")
    else:
        print(f"{file} does not contain the correct columns.")

# Concatenate all valid DataFrames into a single DataFrame
if valid_dataframes:
    combined_df = pd.concat(valid_dataframes, ignore_index=True)
    print("\nCombined DataFrame:")
    print(combined_df)
else:
    print("\nNo valid DataFrames found.")

display(combined_df)

# Shuffe Dataframe

shuffled_df = combined_df.sample(frac=1, random_state=42).reset_index(drop=True)

# 🚧 Save Data to Disk

Let's save all our hard work formatting the dataframe to a CSV for future reference.

- [Pandas DataFrame.to_csv](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.to_csv.html)

shuffled_df.to_csv('data/output/3-merge-data.csv', index=False)
