==> message-categorization/rnn-softmax-multi-class-text-classify/4-train-mnb.ipynb <==
# Overview: Naive Bayes

We'll use [Naive Bayes](https://en.wikipedia.org/wiki/Naive_Bayes_classifier) to build our model.

pip install scikit-learn pandas

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, accuracy_score

# Load the CSV
file_path = 'data/preprocessed-data.csv'
df = pd.read_csv(file_path)

# Handle NaN values: choose one of the following methods
df = df.dropna(subset=['singleMessage'])  # Method 1: Drop rows with NaN values
# df['singleMessage'] = df['singleMessage'].fillna('')  # Method 2: Fill NaN values with a default string

# Extract features and target
X = df['singleMessage']
y = df['reason']

# Split the dataset into training and testing sets (70% training, 30% testing)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

# Preprocess the text data using TfidfVectorizer
vectorizer = TfidfVectorizer()
X_train_vectorized = vectorizer.fit_transform(X_train)
X_test_vectorized = vectorizer.transform(X_test)

# Train a Multinomial Naive Bayes classifier
classifier = MultinomialNB()
classifier.fit(X_train_vectorized, y_train)

# Make predictions on the testing set
y_pred = classifier.predict(X_test_vectorized)

# Evaluate the classifier's performance
print("Classification Report:")
print(classification_report(y_test, y_pred))
print("\nAccuracy:")
print(accuracy_score(y_test, y_pred))


# Example message to predict
new_message = ["Trump will win 2024"]

# Preprocess the new message
new_message_vectorized = vectorizer.transform(new_message)

# Predict the class probabilities
class_probabilities = classifier.predict_proba(new_message_vectorized)

# Sort the class probabilities in descending order along with their corresponding classes
sorted_probabilities = sorted(zip(classifier.classes_, class_probabilities[0]), key=lambda x: x[1], reverse=True)

# Print the sorted class probabilities
for reason, probability in sorted_probabilities:
    print(f"{reason}: {probability:.4f}")
