==> message-categorization/rnn-softmax-multi-class-text-classify/5-inference-rnn.ipynb <==
# Overview

We will load the model and infer data from it.

# Hyperparameters

# Tokenize and pad the text data
max_len = 100  # Maximum length of input sequences

# Load Dependencies

import tensorflow as tf
import pickle
from tensorflow.keras.preprocessing.sequence import pad_sequences

# Load Model and Objects



# Load the trained model
model = tf.keras.models.load_model('models/labeler_model.keras')

# Load other objects
with open('objects/labeler_objects.pkl', 'rb') as f:
    tokenizer, label_encoder = pickle.load(f)

# Prediction

# Example message to predict
new_message = ["evreybody take care! ive been in the sim for a month now. thanks for all the help."]

# Preprocess the new message
new_message_seq = tokenizer.texts_to_sequences(new_message)
new_message_padded = pad_sequences(new_message_seq, maxlen=max_len)

# Predict the class probabilities
class_probabilities = model.predict(new_message_padded)

# Sort the class probabilities in descending order along with their corresponding classes
sorted_probabilities = sorted(zip(label_encoder.classes_, class_probabilities[0]), key=lambda x: x[1], reverse=True)

# Set a threshold to filter out near-zero probabilities
threshold = 0.01

# Print the sorted class probabilities above the threshold
for reason, probability in sorted_probabilities:
    if probability > threshold:
        print(f"{reason}: {probability:.4f}")
